+++ lib/ui/events2/Event2SearchPanel.dart
@@ -214,7 +214,7 @@ class _Event2SearchPanelState extends State<Event2SearchPanel> implements Notifi
       Expanded(child:
         _buildSearchTextField()
       ),
-      _buildSearchImageButton('close',
+      _buildSearchImageButton('close-circle',
         label: Localization().getStringEx('panel.search.button.clear.title', 'Clear'),
         hint: Localization().getStringEx('panel.search.button.clear.hint', ''),
         onTap: _onTapClear,
@@ -291,7 +291,7 @@ class _Event2SearchPanelState extends State<Event2SearchPanel> implements Notifi
           onTap: _onMapView,
         ),
         Visibility(visible: Auth2().account?.isCalendarAdmin ?? false, child:
-          Event2ImageCommandButton('plus-circle',
+          Event2ImageCommandButton(Styles().images.getImage('plus-circle'),
             label: Localization().getStringEx('panel.events2.home.bar.button.create.title', 'Create'),
             hint: Localization().getStringEx('panel.events2.home.bar.button.create.hint', 'Tap to create event'),
             contentPadding: EdgeInsets.only(left: 8, right: 16, top: 12, bottom: 12),