+++ lib/ui/groups/GroupSearchPanel.dart
@@ -134,7 +134,7 @@ class _GroupsSearchPanelState extends State<GroupsSearchPanel>  implements Notif
                       padding: EdgeInsets.all(12),
                       child: GestureDetector(
                         onTap: _onTapClear,
-                        child: Styles().images.getImage('close', excludeFromSemantics: true),
+                        child: Styles().images.getImage('close-circle', excludeFromSemantics: true),
                       ),
                     )
                 ),