+++ lib/utils/AppUtils.dart
@@ -125,7 +125,7 @@ class AppAlert {
                   Navigator.of(context).pop();
                   }, child:
                   Padding(padding: EdgeInsets.all(16), child:
-                    Styles().images.getImage('close', excludeFromSemantics: true)
+                    Styles().images.getImage('close-circle', excludeFromSemantics: true)
                   )
                 )
               )