+++ CHANGELOG.md
@@ -5,10 +5,50 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [5.2.9] - 2024-05-14
+### Added
+- Added Clear button to event filter bar [#4041](https://github.com/rokwire/illinois-app/issues/4041).
+- Created share event set functionality [#4041](https://github.com/rokwire/illinois-app/issues/4041).
+- Added more radio stations support [#4059](https://github.com/rokwire/illinois-app/issues/4059).
+
+## [5.2.8] - 2024-05-10
+### Changed
+- Updated Event info strings for external registration events [#4043](https://github.com/rokwire/illinois-app/issues/4043).
+- Launch wellness daily tip URLs in an external broswer [#4037](https://github.com/rokwire/illinois-app/issues/4037).
+- Hide notifications from Browse and Favorites tabs [#4051](https://github.com/rokwire/illinois-app/issues/4051).
+### Fixed
+- Fixed content attributes selection [#4034](https://github.com/rokwire/illinois-app/issues/4034).
+### Added
+- Added Clear button to image selection widget [#4044](https://github.com/rokwire/illinois-app/issues/4044).
+
+## [5.2.7] - 2024-05-07
+### Fixed
+- Make sure to always return successfully refreshed token in Auth2 service.
+- Fixed miscelanious service dependencies [#447](https://github.com/rokwire/app-flutter-plugin/issues/447).
+
+## [5.2.6] - 2024-04-30
+### Added
+- Perform more precise permissions detection in Auth2Account [#445](https://github.com/rokwire/app-flutter-plugin/issues/445).
+
+## [5.2.5] - 2024-04-26
+### Added
+- Added Group post scheduling [#4027](https://github.com/rokwire/illinois-app/issues/4027).
+### Changed
+- Require privacy level 4 for ability to create groups [#4030](https://github.com/rokwire/illinois-app/issues/4030).
+
+## [5.2.4] - 2024-04-05
+### Changed
+- Group Direct Messages separated from Posts [#4026](https://github.com/rokwire/illinois-app/issues/4026).
+
+## [5.2.3] - 2024-04-02
 ### Changed
 - Acknowledged new application services initialization in parallel [#408](https://github.com/rokwire/app-flutter-plugin/issues/408).
 ### Fixed
 - Fixed local application services dependencies [#408](https://github.com/rokwire/app-flutter-plugin/issues/408).
+- Fixed Survey HeaderBar [#4020](https://github.com/rokwire/illinois-app/issues/4020).
+### Added
+- Added iOS privacy manifest [#4023](https://github.com/rokwire/illinois-app/issues/4023).
 
 ## [5.2.2] - 2024-02-28